compose: Move functions around to reduce circular imports.
