+++ web/tests/compose_validate.test.js
@@ -12,7 +12,6 @@ const {page_params} = require("./lib/zpage_params");
 
 const channel = mock_esm("../src/channel");
 
-const compose_actions = zrequire("compose_actions");
 const compose_banner = zrequire("compose_banner");
 const compose_pm_pill = zrequire("compose_pm_pill");
 const compose_state = zrequire("compose_state");
@@ -148,7 +147,7 @@ test_ui("validate_stream_message_address_info", ({mock_template}) => {
 });
 
 test_ui("validate", ({mock_template}) => {
-    compose_actions.update_placeholder_text = () => {};
+    compose_recipient.update_placeholder_text = () => {};
     compose_recipient.on_compose_select_recipient_update = () => {};
 
     function initialize_pm_pill() {
